daniel revision
