+++ superset/reports/notifications/base.py
@@ -19,24 +19,17 @@ from dataclasses import dataclass
 from typing import Any, List, Optional, Type
 
 import pandas as pd
-from typing_extensions import TypedDict
 
 from superset.reports.models import ReportRecipients, ReportRecipientType
-
-
-class HeaderDataType(TypedDict):
-    notification_format: str
-    owners: List[int]
-    notification_type: str
-    notification_source: Optional[str]
-    chart_id: Optional[int]
-    dashboard_id: Optional[int]
+from superset.utils.core import HeaderDataType
 
 
 @dataclass
 class NotificationContent:
     name: str
-    header_data: HeaderDataType
+    header_data: Optional[
+        HeaderDataType
+    ] = None  # this is optional to account for error states
     csv: Optional[bytes] = None  # bytes for csv file
     screenshots: Optional[List[bytes]] = None  # bytes for a list of screenshots
     text: Optional[str] = None
