+++ superset/reports/notifications/email.py
@@ -27,9 +27,9 @@ from flask_babel import gettext as __
 
 from superset import app
 from superset.reports.models import ReportRecipientType
-from superset.reports.notifications.base import BaseNotification, HeaderDataType
+from superset.reports.notifications.base import BaseNotification
 from superset.reports.notifications.exceptions import NotificationError
-from superset.utils.core import send_email_smtp
+from superset.utils.core import HeaderDataType, send_email_smtp
 from superset.utils.decorators import statsd_gauge
 from superset.utils.urls import modify_url_query
